Add edit test to addRemove.test.js file
