--- src/modules/edit.test.js
@@ -1,3 +1,0 @@
-import {
-  UpdateTask, saveUpdatedTask, UpdateStatus, clearAll,
-} from './edit.js';
